Refactor authentication views and types; update styles and layout variables
- Updated LoginView and RegisterView components for improved user experience.
- Created new AccountView, ProjectView, DataSourcesView, and LabelSchemesView components.
- Removed deprecated backendAuth types and consolidated auth types into auth.ts.
- Adjusted SCSS styles for better spacing and typography consistency.
- Introduced new variables for padding, margin, and font weights in layout and typography SCSS files.
- Enhanced error handling and loading states in various views.
- Updated DataExplorerView and ProjectLabelSchemesView to improve UI responsiveness.
