+++ frontend/src/components/labels/CreateLabelForm.vue
@@ -141,7 +141,7 @@ const handleCancel = () => {
     label {
         display: block;
         margin-bottom: vars.$margin-small;
-        font-weight: vars.$font-weight-heading;
+        font-weight: vars.$font-weight-xlarge;
         color: vars.$theme-text;
     }
