+++ frontend/src/styles/_auth.scss
@@ -92,7 +92,7 @@
         label {
             color: vars.$color-white;
             font-weight: 500;
-            margin-bottom: vars.$margin-tiny;
+            margin-bottom: vars.$margin-xsmall;
             text-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
             display: block;
         }
