+++ frontend/src/styles/variables/_layout.scss
@@ -9,21 +9,21 @@ $spacers: (
 );
 
 // Paddings
-$padding-tiny: 0.25rem; // 4px
+$padding-xsmall: 0.25rem; // 4px
 $padding-small: 0.5rem; // 8px
 $padding-medium: 1rem; // 16px
 $padding-large: 1.5rem; // 24px
 $padding-xlarge: 2rem; // 32px
 
 // Margins
-$margin-tiny: 0.25rem; // 4px
+$margin-xsmall: 0.25rem; // 4px
 $margin-small: 0.5rem; // 8px
 $margin-medium: 1rem; // 16px
 $margin-large: 1.5rem; // 24px
 $margin-xlarge: 2rem; // 32px
 
 // Gaps
-$gap-tiny: 0.25rem; // 4px
+$gap-xsmall: 0.25rem; // 4px
 $gap-small: 0.5rem; // 8px
 $gap-medium: 1rem; // 16px
 $gap-large: 1.5rem; // 24px
